ax_enums.mojom 1
Explicitly casting all uses of enums from ax_enums.idl. This is
preparatory work so we can convert these enums to mojo in a follow-up.

Bug: 650275

+++ chrome/browser/ui/views/toolbar/toolbar_button.cc
@@ -130,8 +130,9 @@ void ToolbarButton::GetAccessibleNodeData(ui::AXNodeData* node_data) {
   node_data->role = ui::AX_ROLE_BUTTON_DROP_DOWN;
   node_data->AddState(ui::AX_STATE_HASPOPUP);
   if (enabled()) {
-    node_data->AddIntAttribute(ui::AX_ATTR_DEFAULT_ACTION_VERB,
-                               ui::AX_DEFAULT_ACTION_VERB_PRESS);
+    node_data->AddIntAttribute(
+        ui::AX_ATTR_DEFAULT_ACTION_VERB,
+        static_cast<int32_t>(ui::AX_DEFAULT_ACTION_VERB_PRESS));
   }
 }
