+++ components/pdf/renderer/pdf_accessibility_tree.cc
@@ -287,7 +287,8 @@ ui::AXNodeData* PdfAccessibilityTree::CreateNode(ui::AXRole role) {
   ui::AXNodeData* node = new ui::AXNodeData();
   node->id = render_accessibility->GenerateAXID();
   node->role = role;
-  node->AddIntAttribute(ui::AX_ATTR_RESTRICTION, ui::AX_RESTRICTION_READ_ONLY);
+  node->AddIntAttribute(ui::AX_ATTR_RESTRICTION,
+                        static_cast<int32_t>(ui::AX_RESTRICTION_READ_ONLY));
 
   // All nodes other than the first one have coordinates relative to
   // the first node.
