+++ ui/views/controls/button/checkbox.cc
@@ -166,14 +166,17 @@ void Checkbox::GetAccessibleNodeData(ui::AXNodeData* node_data) {
   node_data->role = ui::AX_ROLE_CHECK_BOX;
   const ui::AXCheckedState checked_state =
       checked() ? ui::AX_CHECKED_STATE_TRUE : ui::AX_CHECKED_STATE_FALSE;
-  node_data->AddIntAttribute(ui::AX_ATTR_CHECKED_STATE, checked_state);
+  node_data->AddIntAttribute(ui::AX_ATTR_CHECKED_STATE,
+                             static_cast<int32_t>(checked_state));
   if (enabled()) {
     if (checked()) {
-      node_data->AddIntAttribute(ui::AX_ATTR_DEFAULT_ACTION_VERB,
-                                 ui::AX_DEFAULT_ACTION_VERB_UNCHECK);
+      node_data->AddIntAttribute(
+          ui::AX_ATTR_DEFAULT_ACTION_VERB,
+          static_cast<int32_t>(ui::AX_DEFAULT_ACTION_VERB_UNCHECK));
     } else {
-      node_data->AddIntAttribute(ui::AX_ATTR_DEFAULT_ACTION_VERB,
-                                 ui::AX_DEFAULT_ACTION_VERB_CHECK);
+      node_data->AddIntAttribute(
+          ui::AX_ATTR_DEFAULT_ACTION_VERB,
+          static_cast<int32_t>(ui::AX_DEFAULT_ACTION_VERB_CHECK));
     }
   }
 }
